Refactor project package structure
