+++ src/main/java/org/example/config/PluginConfig.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.config;
 
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
