+++ src/main/java/org/example/service/TeamService.java
@@ -1,8 +1,9 @@
-package org.example;
+package org.example.service;
 
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.example.config.PluginConfig;
 import org.jetbrains.annotations.NotNull;
 import java.util.UUID;
 import java.util.List;
