+++ src/main/java/org/example/util/TeamValidator.java
@@ -1,8 +1,9 @@
-package org.example;
+package org.example.util;
 
 import net.kyori.adventure.text.Component;
 import net.kyori.adventure.text.format.NamedTextColor;
 import org.bukkit.entity.Player;
+import org.example.service.TeamService;
 import org.jetbrains.annotations.NotNull;
 
 /**
